+++ source/Enum/05_icontype.ps1
@@ -1,4 +1,4 @@
-enum icontype
+enum notion_icon_type
 # https://developers.notion.com/reference/page#page-object-properties
 
 {